# banking-core/src/main/java/com/alphbank/core/payment/rest/model/Payment.java
package com.alphbank.core.payment.rest.model;

import com.alphbank.core.rest.model.PaymentDTO;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;

import javax.money.MonetaryAmount;
import java.time.LocalDateTime;
import java.util.Locale;
import java.util.UUID;

@Builder
public class Payment {

    private UUID paymentId;
    private UUID fromCustomerId;
    private UUID fromAccountId;
    private MonetaryAmount amount;
    private String recipientIban;
    private UUID recipientAccountId;
    private String messageToSelf;
    private String messageToRecipient;
    private LocalDateTime executedDateTime;

    public PaymentDTO toDTO(){
        return PaymentDTO.builder()
                // TODO
                .build();
    }

}