# banking-core/src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
tags:
  - name: CoreCustomer
    description: Endpoints related to customers
  - name: CoreAccount
    description: Endpoints related to accounts
  - name: CorePayment
    description: Endpoints related to payments
  - name: CoreLoan
    description: Endpoints related to loans

paths:
  /customer:
    post:
      tags:
        - CoreCustomer
      summary: Create a customer in Alph bank
      description: TODO
      operationId: createCustomer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCustomerRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorResponse"
  /customer/{customer-id}:
    get:
      tags:
        - CoreCustomer
      summary: Get a customer by their customerId.
      description: TODO
      operationId: getCustomer
      parameters:
        - name: customer-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorResponse"
    delete:
      tags:
        - CoreCustomer
      summary: Delete a customer by their customerId.
      description: TODO
      operationId: deleteCustomer
      parameters:
        - name: customer-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorResponse"
    patch:
      tags:
        - CoreCustomer
      summary: Update a specific customer's information.
      description: TODO
      operationId: updateCustomer
      parameters:
        - name: customer-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateCustomerRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Customer"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorResponse"
  /customer/search:
    get:
      tags:
        - CoreCustomer
      summary: Fetch all customers.
      description: TODO
      operationId: searchCustomers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Customer"
        "400":
          description: Bad Request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FieldErrorResponse"

  /account:
    post:
      tags:
        - CoreAccount
      summary: Create an account on a specific customer.
      description: TODO
      operationId: createAccount
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
  /account/{account-id}:
    get:
      tags:
        - CoreAccount
      summary: Get an account by its accountId.
      description: TODO
      operationId: getAccount
      parameters:
        - name: account-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
    delete:
      tags:
        - CoreAccount
      summary: Delete an account by its accountId.
      description: TODO
      operationId: deleteAccount
      parameters:
        - name: account-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                type: string
                format: uuid
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
  /account/search:
    get:
      tags:
        - CoreAccount
      summary: Get all accounts by customerId.
      description: TODO
      operationId: searchAccounts
      parameters:
        - name: customer-id
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountSearchResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request

  /payments:
    post:
      tags:
        - CorePayment
      summary: Create a payment. The execution of the payment will depend on the scheduledDateTime field.
      description: TODO
      operationId: createPayment
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
  /payments/{payment-id}:
    get:
      tags:
        - CorePayment
      summary: Get a payment by its paymentId.
      description: TODO
      operationId: getPayment
      parameters:
        - name: payment-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
    delete:
      tags:
        - CorePayment
      summary: Delete a payment by its paymentId.
      description: TODO
      operationId: deletePayment
      parameters:
        - name: payment-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
  /payments/search:
    get:
      tags:
        - CorePayment
      summary: Search payments by various fields.
      description: TODO
      operationId: searchPayments
      parameters:
        - name: from-account-id
          in: query
          required: true
          schema:
            type: string
            format: uuid
        - name: recipient-iban
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"
        "400":
          description: Bad Request

  /loans:
    post:
      tags:
        - CoreLoan
      summary: TODO
      description: TODO
      operationId: createLoan
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLoanRequest"
        required: true
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
  /loan/{loan-id}:
    get:
      tags:
        - CoreLoan
      summary: Get a loan by its loanId.
      description: TODO
      operationId: getLoan
      parameters:
        - name: loan-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Loan"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
    delete:
      tags:
        - CoreLoan
      summary: Delete a loan by its loanId.
      description: TODO
      operationId: deleteLoan
      parameters:
        - name: loan-id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: No Content
          content:
            application/json:
              schema:
                type: string
                format: uuid
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
  /loan/search:
    get:
      tags:
        - CoreLoan
      summary: Search loans by various fields.
      description: TODO
      operationId: searchLoans
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
        - name: accountId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchLoansResponse"
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: string
        "409":
          description: Conflict
          content:
            application/json:
              schema:
                type: string
        "400":
          description: Bad Request
components:
  schemas:
    CreatePaymentRequest:
      type: object
      properties:
        fromCustomerId:
          type: string
          format: uuid
        fromAccountId:
          type: string
          format: uuid
        recipientIban:
          type: string
        amount:
          $ref: "#/components/schemas/MonetaryAmount"
        messageToSelf:
          type: string
        messageToRecipient:
          type: string
        scheduledDateTime:
          type: string
          format: date-time
    MonetaryAmount:
      type: object
      properties:
        amount:
          type: number
          examples:
            - 99.96
        currency:
          type: string
          examples:
            - USD
    Payment:
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        fromCustomerId:
          type: string
          format: uuid
        fromAccountId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MonetaryAmount"
        recipientIban:
          type: string
        recipientAccountId:
          type: string
          format: uuid
        messageToSelf:
          type: string
        messageToRecipient:
          type: string
        executedDateTime:
          type: string
          format: date-time
        scheduledDateTime:
          type: string
          format: date-time
    CreateLoanRequest:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        principal:
          $ref: "#/components/schemas/MonetaryAmount"
        fixedRateInterestAPR:
          type: number
        loanTermMonths:
          type: integer
          format: int32
    Loan:
      type: object
      properties:
        loanId:
          type: string
          format: uuid
        customerId:
          type: string
          format: uuid
        accountId:
          type: string
          format: uuid
        principal:
          $ref: "#/components/schemas/MonetaryAmount"
        fixedRateInterestAPR:
          type: number
        loanTermMonths:
          type: integer
          format: int32
        payoutDateTime:
          type: string
          format: date-time
    Address:
      type: object
      required:
        - streetAddress
        - city
        - country
      properties:
        streetAddress:
          type: string
          examples:
            - Main Street 4
        city:
          type: string
          examples:
            - Stockholm
        country:
          type: string
          examples:
            - Sweden
    CreateCustomerRequest:
      type: object
      required:
        - language
        - countryCode
        - nationalId
        - firstName
        - lastName
        - address
      properties:
        language:
          type: string
          examples:
            - sv
        countryCode:
          type: string
          examples:
            - SE
        nationalId:
          type: string
          examples:
            - 200510312390
        firstName:
          type: string
          examples:
            - John
        lastName:
          type: string
          examples:
            - Doe
        address:
          $ref: "#/components/schemas/Address"
    Customer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        address:
          $ref: "#/components/schemas/Address"
        language:
          type: string
          examples:
            - sv
        countryCode:
          type: string
          examples:
            - SE
        nationalId:
          type: string
          examples:
            - 200510312390
        firstName:
          type: string
          examples:
            - John
        lastName:
          type: string
          examples:
            - Doe
    CreateAccountRequest:
      type: object
      properties:
        customerId:
          type: string
          format: uuid
          examples:
            - 3fa85f64-5717-4562-b3fc-2c963f66afa6
        locale:
          type: string
          examples:
            - sv_SE
        accountName:
          type: string
          examples:
            - Savings account
      required:
        - customerId
        - locale
    Account:
      type: object
      properties:
        id:
          type: string
          format: uuid
        accountName:
          type: string
        balance:
          $ref: "#/components/schemas/MonetaryAmount"
        iban:
          type: string
    UpdateCustomerRequest:
      type: object
      properties:
        address:
          $ref: "#/components/schemas/Address"
        phoneNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
    PaymentSearchRestResponse:
      type: object
      properties:
        payments:
          type: array
          items:
            $ref: "#/components/schemas/PaymentSearchResult"
    PaymentSearchResult:
      type: object
      properties:
        paymentId:
          type: string
          format: uuid
        fromCustomerId:
          type: string
          format: uuid
        fromAccountId:
          type: string
          format: uuid
        amount:
          $ref: "#/components/schemas/MonetaryAmount"
        recipientIban:
          type: string
        recipientAccountId:
          type: string
          format: uuid
        message:
          type: string
        visibleOnAccountDateTime:
          type: string
          format: date-time
    SearchLoansResponse:
      type: object
      properties:
        loans:
          type: array
          items:
            $ref: "#/components/schemas/Loan"
    CustomerSearchResponse:
      type: object
      properties:
        customers:
          type: array
          items:
            $ref: "#/components/schemas/Customer"
    AccountSearchResponse:
      type: object
      properties:
        accounts:
          type: array
          items:
            $ref: "#/components/schemas/Account"
    ErrorResponse:
      type: object
      properties:
        errorCode:
          type: integer
          examples:
            - 4xx
        errorMessage:
          type: string
    FieldErrorResponse:
      type: object
      properties:
        errorCode:
          type: integer
          examples:
            - 404
        errors:
          type: array
          items:
            $ref: "#/components/schemas/FieldErrorEntry"
    FieldErrorEntry:
      type: object
      properties:
        errorCode:
          fieldName: string
          examples:
            - field.nestedField
        errors:
          type: array
          items:
            type: string